Changed galaxy density to a float
Makes a smoother cloud, greater variety in density values, and
eliminates int <--> float conversions.

+++ galacticcluster.h
@@ -5,7 +5,7 @@
 class GalacticCluster
 {
 public:
-    int GalaxyDensity;
+    float GalaxyDensity;
     GalacticCluster();
 };
